pyproject.toml
[tool.poetry]
name = "lms-system"
version = "0.1.0"
description = ""
authors = ["Vadim <vadim.marakushin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
pillow = "^10.3.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
black = "^24.4.2"
isort = "^5.13.2"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
